Moved custom instructions handling to py_methods
